pyproject.toml
[project]
name = "nowlii"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.4"
dependencies = [
    "dj-rest-auth>=5.0.2",
    "django>=6.0.1",
    "django-allauth[socialaccount]>=65.13.1",
    "django-cors-headers>=4.9.0",
    "django-otp>=1.7.0",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.14",
]
